+++ src/main/java/com/appointment/booking/repository/OfferRepository.java
@@ -5,6 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 @Repository
 public interface OfferRepository extends JpaRepository<Offer, Long> {
 
@@ -12,5 +14,5 @@ public interface OfferRepository extends JpaRepository<Offer, Long> {
     String getOfferNameById(Long id);
 
     @Query("select o.price from Offer o where o.id = :id")
-    Double getOfferPriceById(Long id);
+    BigDecimal getOfferPriceById(Long id);
 }
